Fix type errors

+++ polyA/fill_align_matrix.py
@@ -85,7 +85,7 @@ def fill_align_matrix(
         gap_ext = gap_exts[i]
 
         char_complexity_adjustments: Dict[
-            str, int
+            str, float
         ] = calculate_complexity_adjusted_score(
             background_freqs[i], subfam_seq, chrom_seq, lamb
         )
